[6.4.0] build-runfiles: remove temporary file prior to creating it (#19386)
When building with Remote Output Service, bb_clientd has the ability to
restore snapshots of previous bazel-out/ directories. Though the file
contents of these snapshots are identical to what's created in the past,
the files will be read-only. This is because the files may be shared by
multiple snapshots.

We have noticed that most of Bazel is fine with that. Most of the times
Bazel is a good citizen, where it removes any files before recreating
them. We did notice a very rare case where build-runfiles tries to make
in-place modifications to a temporary file that it maintains. This
change ensures that build-runfiles stops doing this.

Closes #19241.

Commit
https://github.com/bazelbuild/bazel/commit/357fc5f81b483713bb7b0319a811c4585975323d

PiperOrigin-RevId: 561615005
Change-Id: I8ca95c7d35df8a53af8f632b10b4a6141d180631

Co-authored-by: Ed Schouten <eschouten@apple.com>