Fix error reporting in java_binary that launcher is not present
It may happen there is a select in use_launcher and then launcher attribute is set to a label flag (not none). While the select may still resolve to False. We can't do any better.

We shouldn't report an error in this case, because the launcher is not used and the user didn't set the launcher attribute.

PiperOrigin-RevId: 562504797
Change-Id: I5a1a020fd4cd1832b600124fa876caf890b0a08b