Don't signal error if the subject for creating connection is already terminated (either with or without error)
otherwise, the error will be sent to the global error handler and be printed which is confusing.

Originally, the code checks with `!hasComplete()` which works fine if there is no error when creating the underlying connection.

With my effort of moving the remote initialization into the background, veryfing the server capabilities will be part of connection creation so it could fail.

Working towards #18607.

PiperOrigin-RevId: 562532842
Change-Id: I5e33d1d9fc62382a77bdcf8617d7561a989bd74b