Add `attrs` to `subrule()`
We apply certain restrictions for subrule attributes:
 - Only private attributes are permitted
 - Computed defaults are forbidden.
 - All attributes must have non-`None` default values

The attribute values are passed to the implementation function as named arguments. In a following change, attributes will be added to the owning rule and values will be fetched from the rule context. That is particularly necessary for `Label`-typed attributes that need to be resolved to `Target`s. We'd also like the attributes to appear in `bazel query` for the owning rule.

PiperOrigin-RevId: 562746241
Change-Id: I4ba492acc5470a1e0e2683403722045e1f21fcb6