Handle the possibility that a dependency is not requested again after a restart.
If this happens, then we need to remove the corresponding reverse dep edge to maintain graph consistency. This change does so by using the same mechanism that removes reverse dep edges when deps change across graph versions.

If all sky functions were pure and hermetic, we would expect dependencies to be stable at a given graph version. However, in practice this is not the case. Input-discovering actions, for example, exhibit unstable dependencies at the same graph version because they use mutable state to track discovered inputs.

PiperOrigin-RevId: 562802006
Change-Id: Ibfd2e4a12a3c093b5a82a2f6eae29cfef567243e