Move the construction of ActionOutputDirectoryHelper into CommandEnvironment.
In a followup CL, the ActionOutputDirectoryHelper will be passed into the prefetcher at construction time. The prefetcher is created by a BlazeModule, and the CommandEnvironment seemed the most appropriate choice out of the set of things BlazeModules have access to.

Also move ActionOutputDirectoryHelper into the actions package, since it's no longer intended to be used solely by Skyframe, and more clearly spell out the guarantees made by its methods.

PiperOrigin-RevId: 563193480
Change-Id: I5387255d22629bdb11b5dd9b65dfc16bb28a4764