Attempt to deflake workspace_resolved_test
I wasn't able to run this test locally, so haven't been able to figure out the exact source of the flakiness. But my theory is that it's flaky because `date +%s` returns seconds since epoch, and the two Bazel invocations might sometimes be so fast that they happen within the same second, cause the flake. So I changed this test to be "definitely nonreproducible" by making it copy a file from outside the repo, which I change between the invocations.

Fixes https://github.com/bazelbuild/bazel/issues/19323

PiperOrigin-RevId: 563351011
Change-Id: I86a8fda441d129306038c15ee99fa63b764b7d4a