Fix parent directory creation when fetching into an externally modified output tree.
The prefetcher currently assumes that the output tree is either empty or consistent with the outcome of the previous build, which isn't true when the filesystem is externally modified or when toggling an output service on and off.

Specifically, we may not assume that the parent path of a build output already exists and is a writable directory, since that precondition is established by action execution (see SkyframeActionExecutor#createOutputDirectories), but a remotely stored output may be prefetched in an incremental build without the respective action reexecuting.

Fixes #19381.

PiperOrigin-RevId: 563369635
Change-Id: Idd521790d4c8ab1b3d807001742ec4eba1730630