Asynchronously verify ServerCapabilities.
Currently, we get and verify ServerCapabilities within `RemoteModule#beforeCommand` which blocks Bazel. However, the result, if not an error, is not needed until execution phrase when Bazel starts to execute actions remotely or query remote cache.

This CL changes Bazel to not wait within `beforeCommand`, but to get/wait the ServerCapabilities when it needs them. This is achieved by moving the process to be a part of channel creation and by using `AsyncTaskCache` to ensure it's only executed once. `beforeCommand` still triggers the start of the process.

Error generated during that process is no longer an `AbruptExitException` but normal `IOException` (and is treated as remote error) which means local fallback works automatically. And in case of building with only remote cache, it also automatically falls back to local execution (instead of exiting with error). As a result, some tests must be updated/removed.

Working towards #18607.

PiperOrigin-RevId: 563399076
Change-Id: I8d4130e81a059b4939a7e745e25abd3916f39f9e