Fix memory regression caused by https://github.com/bazelbuild/bazel/commit/91d2c114df356b70754683482317fa8e15d084b6 .
That change added a dependency on the a nested set that was almost always empty, which resulted an unnecessary and useless Skyframe dependency edge for many actions.

The fix is to check whether the nested set is empty before adding a dependency on it.

Naively, one would think that we could do the same thing with `nonLeaf` above, but `NestedSetBuilder.addTransitive()` already filters out empty nested sets so doing the same emptiness check for `nonLeaf` would be useless.

RELNOTES: None.
PiperOrigin-RevId: 563662609
Change-Id: I4ea3b250f75b9fad8e7532e6e0675469eccb220f