Lift all subrule attributes to a rule that uses it.
We just add the attributes defined on the subrules to the rule class as if they were defined on the rule itself (renaming them to avoid conflict). This seems to be the simplest way to leverage all the existing rule machinery to resolve dependencies *and* get the right results for `bazel query`

A downside of doing it this way, is that it makes these attributes bona fide attributes of the rule. So if we want to ensure these attributes are not accessible via the rule context, it will require performing some extra checks. I haven't done that here and I'm not sure if that is worth it. The subrule proposal is meant to help make the code modular and have better abstractions. If one wishes to go the extra mile to violate that, :shrugs:

PiperOrigin-RevId: 563703792
Change-Id: Ifdb0893208e3693a5cde473f0cfeb3d16583d402