Load Python rules in integration tests instead of using builtins (#2)
This is to facilitate moving the rule implementation out of Bazel itself and into rules_python. This also makes the tests pass when an allowlist is used to enforce that the Python rules are being loaded through rules_python instead of used from builtins.

This the same change as before, but with some internal fixes applied.

PiperOrigin-RevId: 565113515
Change-Id: I999dbfaace8cce2168e718093fe418d548ab1b73

Partial commit for third_party/*, see #19472.

Signed-off-by: Hee Cha <heec@google.com>