[6.4.0] Only use `/showIncludes` if supported (#19521)
The auto-configured Windows toolchain now only enables `/showIncludes`
if the English language pack for VS is installed and the compiler
language can thus be forced to be English. Other languages may use
encodings other than ASCII or UTF-8 and thus fail to be recognized by
Bazel's `ShowInputFilter`. A warning with a suggested fix is shown if
this fails.

See #19451
Fixes #19439

RELNOTES: Compilation actions using the auto-configured MSVC toolchain
are forced to emit error messages in English if the English language
pack for Visual Studio is installed.

Closes #19497.

Commit
https://github.com/bazelbuild/bazel/commit/0f103598bba32439355c94b2c226fb60625c94dd

PiperOrigin-RevId: 565251294
Change-Id: I926c484c21310ee4d7b5e8d5756cb61e45d6e6ac

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>