[6.4.0] Add diff_against_dynamic_baseline option to experimental_output_direc… (#19514)
…tory_naming_scheme.

To support, a new SkyFunction is introduced, BaselineOptionsFunction,
with corresponding SkyValue and SkyKey. Also, a new function
ExecutionTransitionFactory.createTransition was added to immediately
create an ExecutionTransition.

When
--experimental_output_directory_naming_scheme=diff_against_dynamic_baseline,
the behavior is similar to diff_against_baseline, except when "is Exec"
is true (that is, an ExecutionTransition has previously been applied),
the baseline is chosen to be result of the ExecutionTransition being
applied to the normal baseline.

This is meant to resolve an issue with performance when using remote
execution engines that collate requests from multiple users. They were
seeing varying output paths for actions in an execution configuration
depending on how much 'resetting' the exec transition was doing because
of varying user commandlines.

This relates to https://github.com/bazelbuild/bazel/issues/14023 and
https://github.com/bazelbuild/bazel/pull/18002

PiperOrigin-RevId: 537097551
Change-Id: I72966406b7b8af0174a81b638bf0d1fb62466653

Co-authored-by: Googler <twigg@google.com>