Verify `rule` call in `createRule` instead of `export` method
Move verification of attributes and other parameters supplied by rule call into `createRule`. Move all those that don't need to be delayed until the export time.

Main motivation for this change is to have all public attributes in a single location using a single type. This is preparation for `extend_rule` implementation.

Doing verification early is a good practice. There are semantic changes, because there could be a preexisting error in an un-exported rule. Any such unexported rule can't be "fixed" later. So the new semantics helps in early error detection.

Remove state variables in StarlarkRuleFunction. Now StarlarkRuleFunction either has RuleClass.Builder or RuleClass. That's now easier to follow.

Allow Starlark transitions in @builtins defined rules (not @builtins exported rules). That's because this data is available in `createRule`.

PiperOrigin-RevId: 566647536
Change-Id: I7b71a5e47b0ed0d059e66fc5b0b12999ba0c2c6e