Internal change: Fix compliance regression in from unknown commit
I'm not totally happy with this change. We probably should add a method for mock license support.  But this will do to get us on the right track.

The underlying problem is that we have tests which mingle mocks with live rule code. This is fragile. Mock it all, or pull in all of the live things. The first is difficult and may not test reality.  The latter forks blaze and bazel tests.

RELNOTES: None
PiperOrigin-RevId: 566922698
Change-Id: I67ffed21240385c73b56133424a01487cdbe9714