Fix visibility check at definition for aspects
With `--incompatible_visibility_private_attributes_at_definition`, the visibility of aspects (and aspects on aspects) into their implicit dependencies is now checked relative to the aspect's, not the rule's definition.

This requires ensuring the precedence of aspect on aspect attributes over those of the underlying aspect in `DependencyResolutionHelper`, which was hinted at by a comment but not realized.

Closes #19442.

PiperOrigin-RevId: 567333681
Change-Id: I167c45fea8535e0f35c0298ca1cb9f845c9fe2d2
(cherry picked from commit 7a537ca644ac17d565585cd4b3e21bf6243dc001)