Stablize and flip repository_cache_urls_as_default_canonical_id
Introduced in #14268, this flag is very useful for bigger enterprise
context where folks often version bumping dependencies without
remembering to update the SHA256 of the downloaded file, leading to
Bazel picking up older download entries from the repository cache.

As we get more questions about this flag in Slack, marking it as
stable and flip the default seems to be the right move.

Closes #19549.

PiperOrigin-RevId: 567356017
Change-Id: I6402e33f569789545e3ce17ebb42c51a8d56126f