Support rewinding `IncrementalInMemoryNodeEntry`.
`ReverseDepsUtility#getReverseDeps` is renamed `consolidateAndGetReverseDeps` to make it clear that it has a side effect of consolidating the rdep operations. This is critical in the incremental case, as explained by a code comment, and covered by the newly added `IncrementalInMemoryNodeEntryTest#rewindOnIncrementalBuild`.

This change includes only unit test coverage. Basic unit tests for the rewinding lifecycle are moved from `NonIncrementalInMemoryNodeEntryTest` to the shared `InMemoryNodeEntryTest`. More test coverage will be added later.

Note that we still need to respect `--rewind_lost_inputs` on incremental builds before the new behavior can be exercised. For now, an incremental build's `GraphInconsistencyReceiver` still crashes on any inconsistencies.

PiperOrigin-RevId: 567656694
Change-Id: I390a465e79111e837e22bc4fa044b480fb7a0d41