Add support for `use_repo_rule` in MODULE.bazel files
This CL introduces a new directive in the MDOULE.bazel file, `use_repo_rule`, which is a convenient way to declare repos that are only visible to the current module. See https://github.com/bazelbuild/bazel/issues/17141#issuecomment-1622343142 for example usage.

Under the hood, this is implemented as an "innate" module extension for each module that uses `use_repo_rule`. The ID of this extension is a fake one: with the bzl file label of `@@name~version//:MODULE.bazel` and name of `__innate`. Each tag of this extension corresponds to a repo. The name of the tag is the string `${repo_rule_bzl_label}%${repo_rule_name}`, and the attributes of the tag are the attributes of the repo.

Fixes https://github.com/bazelbuild/bazel/issues/17141.

PiperOrigin-RevId: 567722226
Change-Id: I16b8dc562dd20a676118867d831d3b2f5631f767