Automated rollback of commit 4482d6cb5b3b7c248d4b621eb5ab7601c4d18faf.
*** Reason for rollback ***

My understanding was incorrect; uploading the Action message to the CAS is required by the UpdateActionResult spec, see https://github.com/bazelbuild/remote-apis/blob/main/build/bazel/remote/execution/v2/remote_execution.proto#L176.

We should consider a separate code path for the HTTP or disk cache, where we don't have to obey this requirement. But let's just roll back for now.

*** Original change description ***

Stop uploading the Action proto to the CAS.

The Action proto describes the contents of an AC entry. The spec doesn't require it to be also stored in the CAS.

Also rename variables for clarity.

PiperOrigin-RevId: 568141421
Change-Id: I5c9d53f6649f2a1313c5bc03ffd380ccbc6e1859