Implement deriving rule's attribute and properties from parent
Implement rule.parent parameter. I seems nicer and simpler in the implementation space not to follow the design doc. In the design doc, the main argument for a separate call was documentation.

Reuse RuleClass ancestors/parents mechanism to merge attributes and properties from the parent rule. This is the initial implementation where only attributes with different names are merged. If the name matches, the implementation fails.

Change executable and test attributes default to UNBOUND. I think it's safer in extended rule, that the users don't get to set these attributes.

Verify all parameters of the rule call. Here we're very strict on the deprecated parameters, so that we don't need to do more cleanups about them.

PiperOrigin-RevId: 568507009
Change-Id: I9d2027c6edc7f876b9d8f76371e8a51c42a32b1a