Linkstatic support for use_lto_native_object_directory
Adds support for use_lto_native_object_directory when using linkstatic, which allows us to use linkstatic and actions that generate tree artifacts simultaneously.

When use_lto_native_object_directory is enabled, lto artifacts that are not handled by lto indexing are stored in shared.nonlto-obj, instead of shared.nonlto.

LibrariesToLinkCollector currently assumes that they are stored in shared.nonlto, resulting in an error when using linkstatic.

We check if use_lto_native_object_directory is enabled, and if it is, we check if obj files are stored in shared.nonlto-obj instead of shared.nonlto.

PiperOrigin-RevId: 568542260
Change-Id: I53c336ac9fccee2276f28aa46c585d21f478c31b