Fail the build (don't crash) if the starlark exec transition is in error.
Right now this doesn't fail right away, it continues and then fails on patch, so that both errors (creating the transition and applying it) are displayed. We could improve this in the future by failing on transition creation, but for now, unblock further progress by at least displaying a reasonable error without crashing.

PiperOrigin-RevId: 568931663
Change-Id: I7f77addae3f838ad0bc1a778998f6ca09d9d1708