Make package(default_applicable_licenses=...) position independent
This is done by migrating the attribute to use a ComputedDefault instead of reading the partial PackageArgs in Package.Builder. The use of ComputedDefault effectively defers the call to get the default until some time after package loading is done.

RELNOTES: default_applicable_licenses in package() now applies to the entire BUILD file, regardless of where in the BUILD that package() is called.
PiperOrigin-RevId: 568955611
Change-Id: Ida4964d408155a2d019df70b7ae143b4ec9768b5