Consider RCs equivalent to release for `bazel_compatibility`
We want `bazel_compatibiliy = [">=6.4.0"]` to match `6.4.0rc1` so that release candidates are indistinguishable from the actual release via Bazel version compatibility requirements. This allows more realistic testing of releases via RCs without having users to rely on hacks such as `bazel_compatibility = [">=6.3.9999"]`.

Along the way change the error message on incompatibility to include the standard module key identifier, which is `<root>` for the root module instead of `<name>@<version>`, both of which can be empty.

Closes #19678.

PiperOrigin-RevId: 569982709
Change-Id: I7dce89ac1c62c1539d71704f048dca65232c119c