Shell-escape javacopts during detokenization
We expand locations (and hence tokenize) on the javacopts coming via package configuration and the toolchain. So just concatenating the options leads to us dropping escape characters (if any) when we tokenize once more upon flattening the depset. Shell-escaping after the first tokenization avoids losing escaping when we re-tokenize the options just before passing them to JavaBuilder. This also means we should tokenize the command line javacopts once before creating the depset.

Fixes: https://github.com/bazelbuild/bazel/issues/19647
PiperOrigin-RevId: 570029861
Change-Id: Ia6cbac39b52c48776cf50e668a6d9ff072b9c1c3