After https://github.com/bazelbuild/bazel/commit/ceddfb1ece1f8ed7ff81558fa1751e6526df031b, select statements are evaluated before computing and applying rule transition. As discovered, this can lead to issues when those select statements are not resolvable. Currently, this causes bazel to crash which is undesirable as any unresolvable selects should not stop the transition from being executed.
This cl introduces SelectResolutionResult and AttributeResolutionResult which are both @AutoValue or a tagged-union. Each has SUCCESS and FAILURE types that corresponds to successful resolution of select/attribute and failed resolution of select/attribute. A forked method of getAndValidate named getResolvedAttribute is also added. It returns AttributeResolutionResult which is then used in StarlarkRuleTransitionProvider to handle unresolvable selects prior to rule transitions.

PiperOrigin-RevId: 570450889
Change-Id: I32e1e6da3110b042c5791886587ae212ddadf673