Add module_files output group, used to build PCMs from cc_library etc directly
Today, `bazel build :target` on a cc_library will not build its PCM.
To debug problems with module compiles we resort to building a target
that depends on the module. This gets in the way, particularly when a *minimal*
such target doesn't exist and others need to reproduce the failure.

With this change, `bazel build :target --output_groups=module_files` will produce the PCM.

PiperOrigin-RevId: 570629210
Change-Id: Ic07509da42637321a0c65f41b0bba8864e3b6a83