Ensure that the output permissions on a tree artifact's intermediate subdirectories have been set before the prefetchFiles call completes, even in the case where the download occurs in a concurrent call.
Previously, concurrent calls to prefetchFiles would only wait for the downloads to complete, which made it possible for an action to observe the tree artifact before the permissions were set. In practice, a large tree consumed by a large number of actions is required to trigger the race condition, and most sandboxing implementations will mask it, since the real tree artifact subdirectories won't be observed.

Although the output permissions are only a best-effort attempt to prevent unintended modifications, not a watertight guarantee, it's sometimes useful to be able to rely on them, as well as to minimize differences between building with and without the bytes.

To avoid excessive implementation complexity, we continue to make no guarantees about the output permissions when the tree artifact contains nested artifacts, is included in a fileset, or results from an an action template expansion. These are all unusual situations that do not apply to most builds.

PiperOrigin-RevId: 570642032
Change-Id: Id5c2665106d879d0d23dd2446ed8ccb6165941ba