[6.4.0] Fix output materialized as symlink when building without the bytes. (#19739)
This appears to have worked in 6.2.x, got broken in 6.3.x, and is
working again at head. The divergence is too great at this point, so
it's not possible to cherry-pick the relevant changes from head.
Instead, I've written a targeted fix that falls back to downloading
every action output if at least one of the outputs was materialized as a
symlink.

Fixes https://github.com/bazelbuild/bazel/issues/19143.