[6.4.0] Make lockfile's `RepoSpec` attributes more readable (#19748)
String-valued tag and repo rule attributes in the `MODULE.bazel.lock`
file have to be serialized in a way that distinguishes them from
label-valued attributes. This commit uses the fact that the string
representation of labels always starts with `@@` to escape strings only
if they start with `@@` (or happen to be of the form of an escaped
string). Since string-valued attributes rarely start with `@@`, in
particular when specified by humans and not macros, this greatly reduces
the need for escaping in the lockfile.

This commit raises the lockfile version to 3.

Along the way also disables HTML escaping for attributes as it isn't
needed and results in less readable representations of certain
characters.

Closes #19684.

PiperOrigin-RevId: 571037360
Change-Id: I04bb60d371cd09326780004122e729d78c99732a

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>