docs: extend docs for general rule (alias)
Fixing typo in markdown URL and extending the docs on why the `alias` rule may be helpful. I felt it may not be clear how this rule can be useful (based on the description and code example provided).

I was thinking to provide an example of how a `select` could have been used in an alias (i.e. it's convenient to be able to refer to a target by a single label):

```
alias(
    name = "rustfmt_bin",
    actual = select({
        "@rules_rust//rust/platform:osx": "@rust_darwin_x86_64__x86_64-apple-darwin_tools//:rustfmt_bin",
        "@rules_rust//rust/platform:linux": "@rust_linux_x86_64__x86_64-unknown-linux-gnu_tools//:rustfmt_bin",
    }),
)
```

but felt this may be too detailed for the reference docs.

Closes #19763.

PiperOrigin-RevId: 571851166
Change-Id: If647507803f0c94e0e6b49363caa68e73c574338