Check for all fields in `BootclassPathInfo.isEmpty()`
If an instance is supplied to `java_common.compile()` to override the `BootclassPathInfo` used for compilation, it is used iff it is non-empty. Otherwise, compilation defaults to the `BootclassPathInfo` from the toolchain.

Without this change, this has the surprising result that a custom `BootclasspathInfo` gets ignored even if it has `auxiliary` or `system` inputs but an empty `bootclasspath`.

PiperOrigin-RevId: 572322474
Change-Id: Ic74b48d34e3e4d755b36d592bdbbb2f878b20ea1