Treat incompletely rewound nodes like inflight nodes for purposes of invalidation: delete their reverse transitive closure.
Fixes an incrementality bug with rewinding, as illustrated by one of the added test cases. Without deletion of rewound but not yet inflight nodes, bottom-up invalidation won't traverse past such nodes on a subsequent build because the invalidator assumes that any dirty nodes were already invalidated - it does not know how to deal with nodes that are dirty for rewinding.

PiperOrigin-RevId: 572582041
Change-Id: Id5d36735d968d4b0b3199d0fc799e7cc69996b6f