Delete the sandboxfs sandboxing strategy.
This change deletes the strategy with the goal of simplifying the codebase. We will focus our effort instead on improving the other sandboxing strategies to make them faster.

RELNOTES[INC]: The sandboxfs sandboxing strategy is removed. It hadn't been maintained for a long time, it didn't work for most users and it was not consistently faster while being complex to set up. sandboxfs performance is heavily dependent on the specific setup (setup costs are lower, but you have to pay a penalty for the use of each input) and there are scenarios where it is faster and scenarios where it is slower. Overall it is not worth its weight.

PiperOrigin-RevId: 572977096
Change-Id: Idec5ad988bcdcebb4fca48716246d4b1b5ab2984