Make AspectDetails private to AspectsListBuilder
Use AspectListsBuilder in Attribute class instead of ImmutableList of AspectDetails. This makes it possible to hide AspectsDetails in AspectsListBuilder.

The change makes AspectsListsBuilder a principled way to handle a list of aspects, either from command line or from a rule.

The change makes previously immutable Attribute class mutable because of AspectsListBuilder. A followup change will fix this.

The change is a refactor/no-op. There are no semantic changes.

This is a preparation to merge attributes in extended rules.

PiperOrigin-RevId: 573104239
Change-Id: I86d988abc27a0460706c9e078b6cf084604a4ffc