Update toolchain param for the affected action inside proto_common
The current state exposes incompatibility error between exec_group and toolchain param since the exec_group is always an empty one (no toolchain associated to it), while toolchain type can be different to None.

This comes from `incompatible_enable_proto_toolchain_resolution` flag which should be accounted when setting the toolchain param.

Plus, we can't use toolchain_type in an action if it's not registered on the rule, while registration of the toolchain on the rule depends, again, on the `incompatible_enable_proto_toolchain_resolution` flag. This adds one more reason to use this flag also for the `toolchain` param.

PiperOrigin-RevId: 573191399
Change-Id: I42bc25853bd5dc00af8123f1fbab24417f90f9b3