Change default value of experimental_sandbox_async_tree_delete_idle_threads
The value is changed to 4 which means that the symlink trees used by sandboxing
will be deleted  asynchronously in a thread pool that has size 1 when the build
is running and the pool will grow to 4 when the server is idle.

RELNOTES:Symlink trees created by sandboxing will be deleted asynchronously.
Disable behavior with --experimental_sandbox_async_tree_delete_idle_threads=0.
PiperOrigin-RevId: 573290819
Change-Id: Ib344c4fece3bc405dec835ce26dc98edb50feee5