Set the LocalFileType correctly for a symlink in TargetCompleteEvent.
This was an oversight in https://github.com/bazelbuild/bazel/commit/5100d8716f4b9ce25db40c131a94f346055ca9cb. We're also not handling source directories correctly, but given their frowned-upon status, let's ignore it for now.

I have a larger CL to make it impossible to misuse the LocalFile constructor (by passing in the Artifact and deriving the type from it), but it snowballed into a fairly large refactor, so I'd rather do it after Bazel 7 is cut.

Fixes #19151.

PiperOrigin-RevId: 574128307
Change-Id: Ib07150c6b900b1b99e683c8ce7b6cafe6da1500e