Only add symlinked proguard artifacts to filesBuilder when hasProguardSpecs is True.
Some proguard artifacts are always generated in Starlark regardless of whether `proguard_specs` is specified or not because some users use a select to provide an empty list for `proguard_specs` and it's impossible to tell what a select will produce when determining implicit outputs.  When `proguard_specs` turns out to be an empty list, these artifacts should not be requested by filesBuilder, otherwise the build fails.

PiperOrigin-RevId: 574710546
Change-Id: I45ea0144eaaa195e0c1d5f119eabb83aaba48c0f