During aspect evaluation only validate the aspect dependencies
Since there can be conflict between the main aspect attributes names and the names of the rule and the base aspects attributes, revalidating rule owned dependencies while evaluating an aspect on the rule can cause incorrect errors. At the same time, dependencies of base aspects as well as the rule itself are already checked when they are evaluated.

PiperOrigin-RevId: 575355459
Change-Id: I1813de4aca2227d3781c4ab8ff2bf4d05b55d051