Don't follow symlinks when deleting test outputs
Update tools/test/test-setup.sh to not follow symlinks when deleting the
contents of `$TEST_UNDECLARED_OUTPUTS_DIR` after creating the
`$TEST_UNDECLARED_OUTPUTS_ZIP` when `--zip_undeclared_test_outputs` is
enabled.

This fixes a serious bug where an absolute symlink generated in the test
could delete files anywhere on the filesystem.

For example, a `sh_test` containing a line like:

    ln -s "$HOME" "$TEST_UNDECLARED_OUTPUTS_DIR/home"

would have caused the users home directory to be deleted after copying
it in to the output.zip.

Change-Id: Ia4a8a9699e4e2f40498342af55babc5554a9ac93